Cleanup __init__ module
- Expose public API
- Move exceptions to separate module
- Use relative imports
